Move http errors to new seam/errors.rb
